Add db interface with tests and memory backend
This change creates the file db/db.js which contains the interface for
our db operations. The crawler and the dashboard can import this file,
choose a backend, and call the functions that it defines. It will import
the appropriate backend (backend-redis.js for production) and forward
calls.

This change adds a memory backend for rapid prototyping and for testing.

This change also adds some tests (though many more should be added).